+++ src/test/java/strategy/StrategyTests.java
@@ -21,14 +21,14 @@ public class StrategyTests {
         chartBar.close = new double[max];
         chartBar.high = new double[max];
         chartBar.low = new double[max];
-        chartBar.volume = new double[max];
+        chartBar.volume = new long[max];
         for (int i = 0; i < max; i++) {
             chartBar.datetime[i] = "2022-01-" + String.format("%02d", i + 1);
             chartBar.open[i] = RandomUtils.nextDouble(100, 120);
             chartBar.high[i] = chartBar.open[i] + 10;
             chartBar.low[i] = chartBar.open[i] - 5;
             chartBar.close[i] = RandomUtils.nextDouble(chartBar.low[i], chartBar.high[i]);
-            chartBar.volume[i] = RandomUtils.nextDouble(100000, 3000000);
+            chartBar.volume[i] = RandomUtils.nextLong(100000, 3000000);
         }
 
         Action[] actions = strategy.run(chartBar);
